Move implementations of ast_method to pool
