+++ src/ast/pool.c
@@ -91,3 +91,18 @@ void ast_push_vector(AstVectorRef vector, AstRef element) {
   assert(vector && element);
   AST_VECTOR_FUNC(push_back)(vector, element);
 }
+
+
+void ast_initialize_pool(void) {
+  static bool is_first = true;
+  if (is_first) {
+    ast_pool_ctor();
+    ast_string_allocator_ctor();
+    ast_vector_allocator_ctor();
+    is_first = false;
+  }
+}
+
+void ast_finalize_pool(void) {
+  ast_pool_dtor();
+}
