+++ src/ast_method.h
@@ -8,7 +8,4 @@
 #include "ast/external_definitions.h"
 #include "ast/pool.h"
 
-void ast_initialize_pool(void);
-void ast_finalize_pool(void);
-
 #endif  /* KMC_C89_COMPILER_AST_METHOD_H */
